Closes #2
Could not find a way to get the mention names from standard XMPP.
Ended up using v2 of the HipChat API to get user information.

The good news is that v2 of the API no longer requires admin access to get the information.
A normal user token can be used.